+++ exporters/stdout/stdoutlog/record.go
@@ -8,6 +8,7 @@ import (
 	"errors"
 	"time"
 
+	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/log"
 	"go.opentelemetry.io/otel/sdk/instrumentation"
 	sdklog "go.opentelemetry.io/otel/sdk/log"
@@ -15,45 +16,53 @@ import (
 	"go.opentelemetry.io/otel/trace"
 )
 
-func newValue(v log.Value) value {
+func newValue(v attribute.Value) value {
 	return value{Value: v}
 }
 
 type value struct {
-	log.Value
+	attribute.Value
 }
 
-// MarshalJSON implements a custom marshal function to encode log.Value.
+// MarshalJSON implements a custom marshal function to encode attribute.Value.
 func (v value) MarshalJSON() ([]byte, error) {
 	var jsonVal struct {
 		Type  string
 		Value any
 	}
-	jsonVal.Type = v.Kind().String()
+	jsonVal.Type = valueType(v.Value)
 
-	switch v.Kind() {
-	case log.KindString:
+	switch v.Type() {
+	case attribute.STRING:
 		jsonVal.Value = v.AsString()
-	case log.KindInt64:
+	case attribute.INT64:
 		jsonVal.Value = v.AsInt64()
-	case log.KindFloat64:
+	case attribute.FLOAT64:
 		jsonVal.Value = v.AsFloat64()
-	case log.KindBool:
+	case attribute.BOOL:
 		jsonVal.Value = v.AsBool()
-	case log.KindBytes:
-		jsonVal.Value = v.AsBytes()
-	case log.KindMap:
+	case attribute.BYTESLICE:
+		jsonVal.Value = v.AsByteSlice()
+	case attribute.BOOLSLICE:
+		jsonVal.Value = v.AsBoolSlice()
+	case attribute.INT64SLICE:
+		jsonVal.Value = v.AsInt64Slice()
+	case attribute.FLOAT64SLICE:
+		jsonVal.Value = v.AsFloat64Slice()
+	case attribute.STRINGSLICE:
+		jsonVal.Value = v.AsStringSlice()
+	case attribute.MAP:
 		m := v.AsMap()
 		values := make([]keyValue, 0, len(m))
 		for _, kv := range m {
 			values = append(values, keyValue{
-				Key:   kv.Key,
+				Key:   string(kv.Key),
 				Value: newValue(kv.Value),
 			})
 		}
 
 		jsonVal.Value = values
-	case log.KindSlice:
+	case attribute.SLICE:
 		s := v.AsSlice()
 		values := make([]value, 0, len(s))
 		for _, e := range s {
@@ -61,13 +70,44 @@ func (v value) MarshalJSON() ([]byte, error) {
 		}
 
 		jsonVal.Value = values
-	case log.KindEmpty:
+	case attribute.EMPTY:
 		jsonVal.Value = nil
 	default:
-		return nil, errors.New("invalid Kind")
+		return nil, errors.New("invalid attribute.Type")
 	}
 
 	return json.Marshal(jsonVal)
+}
+
+func valueType(v attribute.Value) string {
+	switch v.Type() {
+	case attribute.EMPTY:
+		return "Empty"
+	case attribute.BOOL:
+		return "Bool"
+	case attribute.INT64:
+		return "Int64"
+	case attribute.FLOAT64:
+		return "Float64"
+	case attribute.STRING:
+		return "String"
+	case attribute.BYTESLICE:
+		return "Bytes"
+	case attribute.SLICE:
+		return "Slice"
+	case attribute.MAP:
+		return "Map"
+	case attribute.BOOLSLICE:
+		return "BoolSlice"
+	case attribute.INT64SLICE:
+		return "Int64Slice"
+	case attribute.FLOAT64SLICE:
+		return "Float64Slice"
+	case attribute.STRINGSLICE:
+		return "StringSlice"
+	default:
+		return "Invalid"
+	}
 }
 
 type keyValue struct {
@@ -112,9 +152,9 @@ func (e *Exporter) newRecordJSON(r sdklog.Record) recordJSON {
 		DroppedAttributes: r.DroppedAttributes(),
 	}
 
-	r.WalkAttributes(func(kv log.KeyValue) bool {
+	r.WalkAttributes(func(kv attribute.KeyValue) bool {
 		newRecord.Attributes = append(newRecord.Attributes, keyValue{
-			Key:   kv.Key,
+			Key:   string(kv.Key),
 			Value: newValue(kv.Value),
 		})
 		return true